+ test for App, moves App.test.js to test folder
