+++ potluck-planner/src/App.jsx
@@ -1,7 +1,7 @@
 
 function App() {
   return (
-    <div className="App">
+    <div data-testid='app' className="App">
       App Component
     </div>
   );
