+++ potluck-planner/src/tests/App.test.js
@@ -1,8 +1,8 @@
 import { render, screen } from '@testing-library/react';
-import App from './App';
+import App from '../App';
 
 test('renders learn react link', () => {
   render(<App />);
-  const linkElement = screen.getByText(/learn react/i);
+  const linkElement = screen.getByTestId(/app/i);
   expect(linkElement).toBeInTheDocument();
 });
